Added support for o4-mini model
Both "o4-mini" and "o3" models do not support the temperature parameter. currently only for "o3" model the temperature param was not being added to the chat completions, but it was being added for "o4-mini" model. I made the slight change, to do the same for both. It checks if the model is not ("o3" and "o4-mini"), if it is not then temperature parameter is added, else openai.NOT_GIVEN is passed, as it was done before for "o3" model.